ci: scripts to inject metric urls for ga
